pyproject.toml
[tool.poetry]
name = "sso"
version = "0.1.0"
description = ""
authors = ["Sentium"]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
python-dotenv = "^1.0.0"
cachecontrol = "^0.12.7"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
certifi = "^2022.12.7"
requests = "^2.29.0"
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.6"
SQLAlchemy = "^2.0.12"
pydantic = {extras = ["email"], version = "^2.1.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
fastapi-mail = "^1.4.1"
oauthlib = "^3.2.2"
pyyaml = "^6.0.1"
boto3 = "^1.28.68"
meltano = "^3.1.0"
redis = "^5.0.1"
celery = "^5.3.4"
flower = "^2.0.1"
asyncpg = "^0.29.0"
sentry-sdk = "^1.35.0"
alembic-postgresql-enum = "^1.0.0"
slack-sdk = "^3.26.1"
factory-boy = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
alembic = "^1.10.4"
pytest-mock = "3.10.0"
httpx = "^0.24.0"


[tool.isort]
extend_skip = ["meltano_el", "migrations"]

[tool.black]
extend_exclude = "meltano_el,migrations"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
